fixed a bug

+++ Assignment 2/0112430301_47.c
@@ -8,25 +8,25 @@ different mathematical expressions and outputs the result.*/
 
 int main (void) {
     char choice;
-    float x, y, z;
+    double x, y, z;
     printf("Enter Case (A, B or C): \n");
     scanf("%c", &choice);
     if (choice == 'a' || choice == 'A')
     {   
         printf("Enter three numbers: ");
-        scanf("%f %f %f", &x, &y, &z);
-        printf("Output: %.2f", sqrt(x) + pow(y, 4) + 6 * z);
+        scanf("%lf %lf %lf", &x, &y, &z);
+        printf("Output: %.2lf", sqrt(x) + pow(y, 4) + 6 * z);
     }
     else if (choice == 'b' || choice == 'B')
     {
         printf("Enter three numbers: ");
-        scanf("%f %f %f", &x, &y, &z);    
-        printf("Output: %.2f", (int) x % (int) y / z);
+        scanf("%lf %lf %lf", &x, &y, &z);    
+        printf("Output: %.2lf", (int) x % (int) y / z);
     }
     else if (choice == 'c' || choice == 'C')
     {
         printf("Enter three numbers: ");
-        scanf("%f %f %f", &x, &y, &z);
+        scanf("%lf %lf %lf", &x, &y, &z);
         printf("Output: %c %c %c", (int) x, (int) y, (int) z);
     }
     else
